Split Navigation.swift files
