+++ Sources/TokamakCore/Views/NavigationLink.swift
@@ -1,0 +1,50 @@
+// Copyright 2020 Tokamak contributors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+//  Created by Jed Fox on 06/30/2020.
+//
+
+public struct NavigationLink<Label, Destination>: View where Label: View, Destination: View {
+  let destination: Destination
+  let label: Label
+
+  public init(destination: Destination, @ViewBuilder label: () -> Label) {
+    self.destination = destination
+    self.label = label()
+  }
+
+  /// Creates an instance that presents `destination` when active.
+  // public init(destination: Destination, isActive: Binding<Bool>, @ViewBuilder label: () -> Label)
+
+  /// Creates an instance that presents `destination` when `selection` is set
+  /// to `tag`.
+  //   public init<V>(
+  //    destination: Destination,
+  //    tag: V, selection: Binding<V?>,
+  //    @ViewBuilder label: () -> Label
+  //   ) where V : Hashable
+
+  public var body: Never {
+    neverBody("NavigationLink")
+  }
+}
+
+/// This is a helper class that works around absence of "package private" access control in Swift
+public struct _NavigationLinkProxy<Label, Destination> where Label: View, Destination: View {
+  public let subject: NavigationLink<Label, Destination>
+
+  public init(_ subject: NavigationLink<Label, Destination>) { self.subject = subject }
+
+  public var label: Label { subject.label }
+}
